Add initial files with base functionality
autoplay.py listens to messages on Discord so when a song is added to a queue of a music bot (like Hydra or Chip) it tries to add similiar song. Still doesn't work because it looks like music bots don't listen to other bots messages (only normal users).
youtube.py is file to handle Youtube API to search for similiar songs.